Add shoppinglist-group component, StoreGroup interface
Splitting up the shoppinglist display into multiple components:
The shopping-list will have no logic, just using other component templates.
The shoppinglist-group will be the default view for visually
sorting items in the shopping list by store, this is where we will
later have the ability to remove them from the shopping list.
Next addition will be a shoppinglist-print component that will
be a print friendly display of all of the shoppinglist items.
This one won't use any fancy components, just store title and lists.

the ShoppinglistStoreGroup interface will be how we receive the
shoppinglist, through a GET. The ShoppinglistDisplayItem is
also soley used for display.

The ShoppinglistDatabaseItem is the interface representing how
the database is structured. This one will be used for adding new
items to the shopping list, as well as removing existing ones.