Add routerlink to shoppinglist
in a future branch (time permitting), I will change the hardcoding of
stores on the client to an arraylist of stores which is
populated by finding all of the distinct store values
in the database (products collection/shoppinglist collection)
and making it an api endpoint for the client to call.

Will then use this list to determine product store position in the
array, which will be used in the fragment
to go to the correct store list.

mat-tab-content bodies have id's like
'mat-tab-content-1-0' where the last number
increments to show position in the 'content' array.